pyproject.toml
[project]
name = "safeclaw"
version = "0.1.0"
description = "A sandboxed, policy-driven local dev assistant with least-privilege security, audit logging, and plugin architecture."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["security", "devtools", "cli", "sandbox", "ai-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5.0",
    "pytest-cov>=5.0",
]
planner = [
    "httpx>=0.27.0",
]
dashboard = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
]

[project.scripts]
safeclaw = "safeclaw.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.setuptools.packages.find]
include = ["safeclaw*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
